Stopped tmwserv from writing to its configuration file
This was confusing, since changes made to the config file while the
server is running are lost once the server exists. Also, XML comments
were being stripped by the server.

The command line option to set the port now only applies to a single
run, and isn't saved in the configuration. There shouldn't be a need for
the server to write to its configuration file, it might not even have
the rights to do so.