Added the initial version of a server monitor
It launches the account and game servers and restarts them when they
quit unexpectedly. When a core file is found after a server quit, a
backtrace is emailed when the CRASH_REPORT_RECEIVER environment variable
is set.

The monitor should be terminated with the TERM signal, which makes sure
the servers get some time to shutdown gracefully.