Don't return configuration value by reference
The lifetime of the default value is often not long enough to survive to
the next statement, so returning it by reference is dangereous.

At the moment there may not be any place that actually tries to use the
returned value as a reference, but it's better not to wait on it to
happen. This was already fixed on the client after it lead to strange
crashes.

Reviewed-by: Jared Adams