Modified the npc_trade() lua function to permit selling the whole player inventory.
It permits to open a sell box with every items in the player inventory
as requested by Striker.

Also added different return value support to both the buy selling functions,
and made fixes where relevant.

The test.lua script will be upgraded to show examples in a next commit.

What's left to be done is to fix the inventory handling for both selling
functions. (Sigh...)

Concerns: Manasource mantis: #78, #101.
Reviewed-by: Jaxad0127