Changed the location of configuration, logs and stats
Instead of searching for the configuration file in ~/.manaserv.xml, the
file is now expected to be in the working directory of the server. The
logs and statistics will also be written there.

This should make it easier to run differently configured servers on the
same machine, and should also be a bit more straight-forward to setup.

Reviewed-by: Yohann Ferreira