Small fixes.
There is no longer any need to send attribute change information for
character or correction point changes to the account server. This is
now handled elsewhere in a dedicated function.
The account server was already doing this but the game server hadn't
been updated. This wasn't causing any major problems since this data
was being sent at the end of the packet.
Update documentation for the functions accordingly.

Also adjust a misplaced opening curly brace in attribute calculation
which could cause modifiers to be applied wrongly in certain cases.

Reviewed-by: Kage