SQLite: Do a proper upgrade of the mana_characters table
We can't leave the obsolete columns around since they are marked as
NOT NULL, so insertion to the table fails if no values are provided for
them.

With SQLite, our only option is to create the table and copy the data
over, which is what the update script now does.

The script was modified so that it is fine to run it again on a database
that was already updated to version 11 before.