Use a transaction when handling a GAMSG_PLAYER_SYNC message
This message can contain a lot of small database updates, which at least
on my system are way more efficient when performed in a transaction (now
it takes no more than 1 second vs. about 14 seconds before). Not saying
this is normal, my guess is that it's due to using full partition
encryption.

I've also prevented the thing from entering an infinite loop in the case
of a wrong message, and corrected some variable names.