Fixed unregistering on the server-side.
When registering or logging, The client is hashing the password
for sending it safely. And the server is hashing it also
to store it the same way. Hence, the password ends hashed twice,
which is correct because the server can't trust the client anyway.

At unregister attempt, the server wasn't hashing the password
before comparing it.

Also while on it, I made the corresponding SQL query use the try catch
method and only delete the account in memory when it's also done
on the Db.

Reviewed-by: thorbjorn, Freeyorp.