Implemented @die chat command
The @die chat command kills the evoking character by setting its hit
points to 0. The example permissions.xml assigns it to the "tester"
rights group and not "players" because server admins who are not aware
of this command might run into trouble when they assume that players
don't have a method to teleport to their spawn point voluntarily.

Resolves: mantis ticket #192

Reviewed-by: Jaxad0127