Transaction logging of various command, new transaction codes and style fix.
The @mute, @goto, @attribute and @announce commands now logs transactions
to the database.

Added new transaction codes to database. I've preliminarily added a lot
of codes for not yet supported commands so that we needn't do a database
update with every single commit which implements one.

Also using the equivalent "say" helper function in place of the
"GameState::sayTo" method in the command handler.

Reviewed-by: Kage