Add an heartbeat time flag in the statistics file.
This way, any kind of external service can know about the running
state of each servers, even if only the account server is
updating the statistics because:

1. When the account server is running, each <gameserver> tag is
a running game server. Otherwise, it's not listed.
2. Whenever the account server stops running, the stat file
isn't updated anymore and so the <heartbeat> tag isn't.
3. Game servers without any contact with an account server will
try to reconnect to the account server and kick out every players
so they aren't considered as running in that case.

Reviewed-by: Crush.

Resolves: Mana-mantis #270.