Fixed crash on first map update
The crash happened when it was trying to move an NPC from its old zone
to its new zone. The old zone was based on the old position, which was
(0,0). That crashed since its zone didn't match its old position, but
its new one.

This fix makes sure to update the old position to be in sync with the
zone, by resetting it after being inserted into the MapComposite.

Reviewed-by: Yohann Ferreira