Path finding: Moved F cost from PathInfo to Location
The F cost represents the estimate of the total cost from source to
destination. It is only relevant to sort the locations to be checked, so
there is no point in assigning it to each coordinate of the map.

With maps up to 200x200, the memory usage is reduced by 160 kB! :P

Reviewed-by: Bertram