Improved error handling in chr_shake_screen Lua function
It's better to use luaL_checkint than manually checking and printing out
an error, since using the helper function makes a more informative error
message.

Also, scripts errors need to be raised before constructing objects,
since due to the way Lua does error handling the destructors of the
objects will not be called.

Also fixed an issue with the 4th parameter, which was first cast to
integer and then multiplied by 10000. Now the multiplication happens
before the casting to integer.

Reviewed-by: Crush