Fixed problems with loading XML files containing Windows newlines
By using xmlParseFile instead of xmlParseMemory, the system-dependent
newlines should be handled automatically.

The .tmx.gz files should still be supported, but instead of manually
decompressing them the xmlParseFile function should take care of that.

It also fixes the leaking of XML documents in both the SkillManager as
well as the PermissionManager, since they now rely on XML::Document,
which cleans up automatically.

Reviewed-by: Stefan Dombrowski