Initialize attribute base and cached modified values
Attributes that didn't get a value yet are not communicated from account
to game server. This could leave some attributes with uninitialized
(random) values, like the amount of gold or some uncalculated attack
speed bonus.

Defaulting the value to 0 seems to be appropriate for all current cases.
When another default value should be used, it can be specified in
attributes.xml and will be sent from the account server.

Reviewed-by: Stefan Dombrowski