Removed inappropriate assignment to mOld in Being::findPath
Being::findPath doesn't change the position of a being, and so should
not be syncing mOld to the current position. When invoked on the wrong
moment, this could cause MapComposite::update to fail to realize that a
being has moved into another zone.

Also removed some other usages of mOld that were not necessary, to make
its purpose clearer and make a potential cleanup easier.

Reviewed-by: Yohann Ferreira