Added a simple crafting system
A client can craft something using the @craft command.

The command needs a list of item names and amounts. The gameserver checks
if the character has these items in the inventory and then passes the list
together with the character handle to the lua script function on_craft in
the script file scripts/crafting.lua.

This function can then be used to evaluate if the list is a valid crafting
combination and when this is the case take or give items.

Implemented two example crafting scripts there, one which enforces exact
item order and amount and one which doesn't. Both are disabled per default
and one needs to be enabled by uncommenting a line.

Also gave the player group permission to use the @craft command in
permissions.xml and added two new items (wood and iron) required for the
example crafting combination.

Resolves: #333

Reviewed-by: bcs86, Bertram