Fixed autoattacks basic handling.
The server is now using the autoattacks system to perform damage
based on the attack's timer.

I also added a default bare knuckle attack when the character
is unequipped.

As a result, maggots can be killed again with bare hands now.

Known issues left:
This isn't tested against equipment changes for now, and the client
isn't in sync with the attacks speed.

Reviewed-by: Crush.