Changed ItemEquipInfo to only be a pair of unsigned.
Previously it was a list of pair ofunsigned int,
used to get the equipment slot requirements.
It didn't make sense as the ItemEquipsInfo was already a list of it.
So you had to manage with a list of list and that ended up being
irrelevant to the need.

I had to disable the equip function to permit compilation.

This will be fixed in my next commit.