Basically redid equip and unequip functions().
I made the system handle the fact that equipment item
are completely unlinked to the inventory items.
Equip items now have a unique itemInstance number permitting
to equip the same item type multiple time when the slot capacity
is wide enough to do so.

I also prepared the functions to welcome in the near tests against
scripted equipment.

The equip process is known to be working server-side but the unequip
process has yet to be reviewed, even if implemented.