Prevent server crash by not having beings talk after spawn for now
At the moment it is the responsibility of the script to make sure any
references to beings passed into script functions are valid.

This means you can't schedule delayed scripts like the one making
maggots say 'Roaaarrrr!!!', since the being might have been removed
before the script gets executed. In the case of this bug the maggots are
removed by some script code testing 'mana.monster_remove'.

We should of course fix the way actor handles are used in Lua so that
scripts can't end up crashing the server.

Mantis-issue: 384
Reviewed-by: Yohann Ferreira