Return a const & from NameMap to avoid potential deep copies
When I introduced NameMap it was only used with pointers, but now it is
also used with std::string so it probably makes sense to keep a default-
constructed value around so that a reference can be returned rather than
a copy.

NameMap::find was renamed to NameMap::value to make it more clear that
it doesn't return an iterator, like std::map::find.

Reviewed-by: Yohann Ferreira
Reviewed-by: Ben Longbons