Removed the inventory check
We can't arbitrarily remove inventory items on initialization, and
shouldn't have to, because what's in the DB was correct when it was
stored.

Also, there should be no need to check the inventory after inserting
items, since the insert method should already make sure not to insert
items when there is no space.

Reviewed-by: Yohann Ferreira