Merged three global script states into one
These scripts could trivially share one script state, since the methods
called on them from the server are not overlapping. This does leave them
open to access each other's global variables, but that's the problem
with global variables.

The one remaining global script file name is now configurable, so that
it may also be set to a script in a different scripting language. The
two related script options are:

  script_mainFile (default: scripts/main.lua)
  script_defaultEngine (default: lua) - renamed from defaultScriptEngine

Reviewed-by: jurkan
Reviewed-by: Yohann Ferreira