Made the game server execute the chr_respawn_accept script even in case of disconnection.
I made the Character::disconnected() function handle that case,
permitting also to centralize GameState::remove() calls there.

I also made the GameState::enqueueWarp() function test whether
the Character pointer is about to be deleted, so that the warp
can be handled directly to avoid a crash.

Last but not least, I also made the Character::update() function
not update the Character specials and hp to avoid discrepancies
seen in the client.

Resolves: Mana-Mantis #309.

Reviewed-by: Ablu.