SQLite: Added a timeout for trying to acquire a database lock
Without a timeout, trying to write to a locked database fails immediately.
Now it will only fail when it can't retrieve a write lock within a second.

Still, we need to go through all the SQLite code some time, since there
is a lot of error handling missing. We also need to make sure that errors
are handled correctly on the higher level.

Reviewed-by: Yohann Ferreira