Made some global Lua variables local
We have to be careful with introducing globals now that there is only a
single Lua state, so we shouldn't use globals unnecessarily. Any variable
should be declared 'local' unless there is a reason to make it global.

For additional safety we can also think about disallowing the use of globals
entirely. This also helps to catch typos in scripts.

Reviewed-by: Erik Schilling