Removed the last direct call to global script function
The ScriptAction of the TriggerArea (which can be created by
mana.trigger_create) was still using a named global function for its
callback. Now it also uses a reference to a script function.

Since it was the last occurrence of a call to a global script function,
I've also removed the Script::prepare(std::string) overload.

Reviewed-by: Erik Schilling
Mantis-issue: 299