Fixed problems with map-bound world state variables
Due to a wrong primary key, which covered only the state name, it was
impossible to use the same state name on different maps. This has now
been fixed.

Another problem was that the map variables were being included in the
global variables, because the related database query did not filter on
the map_id column properly. While fixing that, the map_id column now
allows explicitly marking a state variable as global (with the value 0)
or system variables (with the value -1).

System variables are currently not accessible from scripts, but that
could be changed later.

Reviewed-by: Yohann Ferreira
Reviewed-by: Erik Schilling