Changed Item to a component of Actor
Items also have positions, so the ItemComponent only makes sense as part of
an Actor. Later on it will probably be part of an entity that also has an
ActorComponent.

Since it was annoying to update all the places where items were created,
I've introduced a function for this.

The component types are now prefixed with "CT_" because I wanted to introduce
an 'Item' namespace which would otherwise be conflicting. The component types
enum isn't used much in the code so it can look a bit ugly.

Reviewed-by: Yohann Ferreira