Fixed Lua error reporting for NPC scripts
This was due to a naming conflict between the 'debug' function I had recently
introduced and the 'debug' library provided by Lua. This caused problems when
trying to use debug.traceback for printing a backtrace of the error.

There's multiple ways to avoid the naming conflict. I opted for writing the
related helper functions in all-caps.

Also added an ERROR log function now. As Erik pointed out, there is no conflict
anymore with Lua's 'error' function with the new naming style.

Reviewed-by: Erik Schilling